Created data normalizer with rate of return and volume taken into account. Regression Experiment now can process a true experiment.

+++ pysandbox/minerva/features/feature_generator.py
@@ -28,7 +28,7 @@ class FeatureGenerator(object):
         '''
         Constructor
         '''
-        self.datas = np.array(datas, dtype='float')
+        self.datas = np.array(datas)
         self.generators = generators
         self.generator_params = gen_params
